HIL/requirements.txt
pytest
pyserial
aprs
pigpio
